Removed all melee attack code and refactored code into fewer classes.
