+++ core/src/networking/threads/BasicThread.java
@@ -1,0 +1,30 @@
+package networking.threads;
+
+import engine.Config;
+
+import java.io.IOException;
+import java.net.DatagramPacket;
+import java.net.DatagramSocket;
+import java.net.InetAddress;
+
+public abstract class BasicThread extends Thread {
+
+    protected DatagramSocket socket;
+
+    @SuppressWarnings("InfiniteLoopStatement")
+    @Override
+    public void run() {
+        while (true) {
+            byte[] data = new byte[Config.PACKET_SIZE];
+            DatagramPacket packet = new DatagramPacket(data, data.length);
+            try {
+                socket.receive(packet); // Warning: will wait indefinitely
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            this.parsePacket(packet.getData(), packet.getAddress(), packet.getPort());
+        }
+    }
+
+    protected abstract void parsePacket(byte[] data, InetAddress address, int port);
+}
